split workspace index limits into dedicated microcrate
