+++ Cargo.toml
@@ -17,6 +17,7 @@ members = [
     "crates/perl-parser-pest",
     "crates/perl-semantic-analyzer",
     "crates/perl-workspace-index",
+    "crates/perl-workspace-index-limits",
     "crates/perl-workspace-index-slo",
     "crates/perl-refactoring",
     "crates/perl-incremental-parsing",
@@ -130,6 +131,7 @@ allow = [
 "perl-symbol-table",
     "perl-uri",
     "perl-workspace-index-slo",
+    "perl-workspace-index-limits",
 
     # Tier 3 — Analysis and indexing
     "perl-workspace-index",
@@ -298,6 +300,7 @@ perl-diagnostics-codes = { path = "crates/perl-diagnostics-codes", version = "0.
 
 # Tier 3: Two-level dependencies
 perl-workspace-index = { path = "crates/perl-workspace-index", version = "0.10.0" }
+perl-workspace-index-limits = { path = "crates/perl-workspace-index-limits", version = "0.10.0" }
 perl-workspace-index-slo = { path = "crates/perl-workspace-index-slo", version = "0.10.0" }
 perl-incremental-parsing = { path = "crates/perl-incremental-parsing", version = "0.10.0" }
 perl-refactoring = { path = "crates/perl-refactoring", version = "0.10.0" }
