+++ crates/perl-workspace-index-limits/README.md
@@ -1,0 +1,3 @@
+# perl-workspace-index-limits
+
+Shared resource and performance limit types for `perl-workspace-index`.
